// src/types.ts
export enum OrderType {
	GTC = "GTC",
	FOK = "FOK",
	GTD = "GTD",
	FAK = "FAK",
}
// ...
export enum Chain {
	POLYGON = 137,
	AMOY = 80002,
}
// ...
export type TickSize = "0.1" | "0.01" | "0.001" | "0.0001";